feat: mock control program communication (#45)
`Score#result` is changed from `Int` to `Long` because Firestore only stores 64-bit numbers.

+++ src/main/kotlin/com/github/frozensync/tournament/Score.kt
@@ -2,5 +2,5 @@ package com.github.frozensync.tournament
 
 // other fields left out for now for easier testing
 data class Score(
-    val result: Int
+    val result: Long
 )
